Huge commit: Now committing entirely in using ReactJS.
This commit needs some breakdown.
1. All UI elements now use React Components. The only place where
we use Blaze Template is in base.html, and even then it's just specifying
which React Component to display.

2. All routing now directs to rendering in React. Unfortunately Iron Router/react-meteor
doesn't properly handle route change, as Components do not unmount properly. To fix this,
we call render from React directly, and manually unmount the component every time the route
changes. To do this, we create a special div called 'ReactComponentDiv' which houses all React
Components.

To accomplish this, we need to expose the Components as global objects (until we have a better
dependency management system.) To better decouple the components, I created a *PageContainer
type component where it strictly renders all the contents of a page.

3. Implemented login Logout buttons.

4. Implemented NavBar and polished UI.