Implemented management table for Judges and Rooms.
Part of issue: #46

Abstracted the logic for transformation into the respective contexts.

Initialize or load Hot table depending on what context we're in.

Currently, implementation for rooms is rather crude. It always registers the entire room,
with no specific methods for update and delete.

TODO: Better deal with duplicate rooms.