Removed initial registration table and substitute with enable/disable checkbox.
1. Instead of the user choosing if they want to enable public registration up front,
we now have a slider checkbox in the management page to determine if public
regitration is open or not. This makes more sense.

As a consequence of this, several components are now deleted. Like the chooser container,
and the registration table. The logic to check those are removed as well.

The Tournament registration component is still there, though.

2. New method is added to toggle enable public registration. And the existig method to enable
public registration is deleted.