Initial implementation for checking invariants.
Issue: #48
Invariants checked are __initialized__ whenever  a tournament or round is created.

However, the invariant checks are wrapped in a Tracker.autorun function, and it is
dependent on the Tournaments Meteor Collection as a reactive data source.
This practically means that invariants are always checked  whenever a tournament is
modified in any way.

Note that the current implementation just sends violations as strings. I think it's much
better now to properly implement a globals structure, and wrap those strings in objects
with global keys. This allows the client to properly check and deal with invariant violations.

TODO: Whatever that was in the last paragraph. Surface invariant violations to the UI and
hook up the logic appropriately.