Change invariant checks to be manually added after each method calls.
Since Meteor doesn't support autorun in the Server, we're currently stuck with
manually calling our invariant checks in every method calls.

Also, added types in InvariantChecker.js

Added a new tournament schema for currentInvariantViolations as Session
is only supported on the client, and also it makes more sense as well because
the invariants are localized to the tournaments.