Implemented surfacing violations to UI. Other changes as well.
Part of issue: #48

1. In Management.jsx, have a new warning message DOM that lists all the violation messages
in the tournament.

Initial implementation to disable the create round button if the last round is not finished. Once we
make the validator function extensible, this code should invoke and consume the result there.

2. In InvariantChecker.js: fixed bug, added checks for enough judges.

3. In AfterMethodsFunctions.js, return the violation object instead of just the string key. We need the message.

4. In RegistrationMethods.js: Make sure that we update the items rounds details accordingly when they are
added.

5. In Schemas.js: Put blackbox as true on Object type (we need to clean up false validates etc after this)
and added a validator function.

Ideally, the validator function should live somewhere else, and should be extensible so it should be able to be
called by the client as well. This makes sure that we use the same checks, and the client can use
the result of the checks to show various UI components as disabled/enabled etc.