Initial implementation of Drag and Drop for Judges in assigned view. Other improvements.
Big commit, part of issue #53 and #52

1. In Management.jsx:
 - Implemented drag and drop using old school mouse events. Decided to use this
because HTML5 feels clunky, and react-dnd is difficult to integrate within Meteor.

Created a higher-order component called connectDraggable and connectDroppable
to wrap an existing component, giving it  drag and drop capabilities. The state  is then
bubbled up to the parent to be handled. Idea is based on here:
https://medium.com/@dan_abramov/mixins-are-dead-long-live-higher-order-components-94a0d2f9e750

Eventually we'll need to refactor out this code from Management.jsx, as the class file
is getting too huge. I'm a speed demon currently, though.

2. Added a SchemaHelpers.js to help with schema-related functions.

3. For Invariants check, currently we'll limit it to checking it everytime we update something
after we create a round, and stop it once we assign the round because the invariant
check is useless. Seems to greatly improve speed.

More to come.