Implemented algorithm for assigning roles properly to teams in round 2 and above.
AssignmentAlgorithm is no longer a static helper class. It is now a dynamic class and has
to be instantiated first. A synchronous assign() call is made to assign everything, before
its getters can be called.

This has to be done because we need to assign teams as well when we're assigning rooms,
which makes sense because the teams' role depends on room assignment.

For the assignment algorithm itself, performance is sacrificed for clearer, more concise code.
Performance doesn't seem to be a bottleneck for now.

Small changes to the UI to show the total points of each team.