Implemented a better strategy for judge allocation.
The current algorithm is very simplistic, it just goes through
the list of judges and sorts them in increasing order of importance:
average rank -> no . of judges in room -> odd number of judges in the room
-> no institutional collide -> empty rooms must be filled first.

The Chair is then given to whichever judge that has the highest rank in the
room.

The problem with the current implementation is that it doesn't
take into account of possible 'exchanges' after the initial allotment.
While it is possible to implement that, I don't see a good reason to do
it for now until we are more certain about what's most important
when allocating judges.