JSHint Compliance
For all of our private methods in AssignmentAlgorithm.js, we tell
JSHint that our 'this' is valid.

Also, for now we use /*jshint loopfunc: true */ , I know it's problematic in
terms of performance and efficiency, but since we're not using a
traditional for-loop, the risk of a problem is lessened.