Remove isResultPublishable from the RoundSchema
For now, we have a tournament define a 'enablePublicView' property to
determine if all information should be published. In the near future,
we might want to add more fine-grained properties such as whether round
results should be published, and separate it from other information
such as a news feed for example.

However, we likely will not need round-level permissions for now.