Added manual test scripts. TEMPORARY.
A problem with the test script is that it's loaded on the client side
for now, which is not ideal, even though we do a production check. Not
sure how we only want to load this client side script in a production
environment.

Perhaps using a router would work, needs more investigation.