Implemented better row removal.
1. If there is a round that is not in initial state, we do not allow
removing of rows. This may not be optimal for judges/rooms, but we'll
roll with this for now.

2. We allow removing of rows if a user selects the entire row (multiple
rows included) and hits delete, which will cause the change to be an
empty string. (Any changes that cause ALL changes to empty string will
be considered as deleting a row, as long as the entire row is selected).

3. Our existing implementation of the dialog confirmation popup works
nicely here. :)