Fix bugs with delete round.
1. update publicRounds accordingly when deleting round. Doesn't matter
if it cannot find the publicRounds, Mongo will not complain, just
silently not delete.

2. When deleting round, switch the context back to team management. Use
the proper pattern of bubbling method call to the parent.

TODO: Change all Session-based state to React states. It's easier to
reason with things, though I'm already using the React state pattern
for all my components.