Better checks on room updates.
1. In ValidatorHelper.js, check for duplicate results as part of room
score add up check. We allow duplicate results when changing team
results, not when determining if it's ok to finalize the round.

2. In TournamentManagementContainer.jsx:
 - Got rid of the useless date checks, that never worked in the first
place.
 - Now checks if all results are the same before rendering.
 - Since we want to set some local state in a room, we transform our
room props to a state. This allows us to set the state once we change a
result, locally. This is a working caching solution.
 -  We also show the _actual_ result/score next to the team/debaters
name, because now we've enabled live data changing. Which is cool. We
do this because if we want to mutate the input text box itself, we'll
need multiple refs. Putting it in a readonly label gives it to us for
free, albeit with a slightly worse UX.