Updated to Semantic-UI 2.0
A few bugs abound, mobile view is still not a solved problem.

Bug 1: The toggle switch doesn't change color back when re-enabled if
click on the toggle itself. But will switch if click on the text.

Bug 2: We might have a performance problem.

Also, set back Schemas.js to server because it's screwing up our
performance.