Use meteor-useraccounts for user accounts.
Part of issue: #67

1. Use meteor-useraccounts for our admin accounts workflow. It provides
basic functionality such as email verification (not implemented for
now) and forgot password (as easy as a settings toggle).

We still have problems with routing, but it should be good for now.

The way we implement this is by creating a Blaze Template (only
supported way of doing it for now) and route to it appropriately.

2. Due to us mixing React and Blaze Templates now, change how we
structure our grid between pages.