Remove use of props in setting initial state.
Using props in initial state is an anti-pattern. When we need to get a
variable from a prop, we do it inside the render call.

This fixes the problem where a prop may be changed, but the component
just updates and never sets the state to the new prop.

The only exception to this is in ActiveRoomsComponent, because we're
doing some client-side magic and (hopefully) be very very careful about
how we manage our state.