Implemented public teams/judges.
Part of issue: #68

The best way to go about it would probably be to refactor our data
schema to contain round information only in rounds. But that needs to
be done in the future.

Second best is to update publicTeams/Judges whenever the actual ones
change, but that takes too much of a performance hit.

So we get a reasonable compromise by updating them whenever a round is
activated.